+++ back/batch/src/main/kotlin/com/metamong/util/apartment/AreaConverter.kt
@@ -6,15 +6,15 @@ import java.math.RoundingMode
 object AreaConverter {
     private val PYEONG_DIVISOR = BigDecimal("3.3058")
 
-    fun toPyeong(squareMeter: BigDecimal?): Short? {
+    fun toPyeong(squareMeter: BigDecimal?): Int? {
         if (squareMeter == null || squareMeter <= BigDecimal.ZERO) return null
 
         return squareMeter
             .divide(PYEONG_DIVISOR, 0, RoundingMode.HALF_UP)
-            .toShort()
+            .toInt()
     }
 
-    fun toPyeong(squareMeterStr: String?): Short? {
+    fun toPyeong(squareMeterStr: String?): Int? {
         if (squareMeterStr.isNullOrBlank()) return null
 
         val squareMeter = squareMeterStr.trim().toBigDecimalOrNull() ?: return null
